Prevent arrow and space keys from scrolling
When the exported game is contained in a page larger than the current
viewport (such as when hosted on Kongregate), the arrow keys and
space bar cause the page to scroll. To avoid this, prevent propagation
of these events.

The default behaviour must be prevented at the beginning of onKeyDown
because any key in the repeat key buffer will not be checked by
checkKey (so holding down keys would cause it to scroll).