Modified logging to append cache as child
Method:
* Create a div per cache dump, give it a unique identifier and set it's
innerHTML as the cache contents.
* Append the div as a child to consoletextarea.

Massive performance boost. Apparently a lot of the work of the
.innerHTML += x is actually destroying all the elements before
recreating them again. I profiled this against PSTG with death disabled
and I went from 4.3mins total to 39s total with verbose logging. In
fact, as far as I can tell, the performance penalty of having verbose
logging enabled is now negligible. Even taking out the times repeated
summarisation was only 49s total, and this is compared with 4.3mins for
the summarised .innerHTML += version.