Support matching on ellipsis.
Ellipsis no longer needs to be alone; it just needs to be last.  The
ellipsis will only span over matching blocks.  This is useful for
line-of-sight sorts of things. For example, the rule for the Wizard in
Heroes of Sokoban can be simplified to:

    [ > Wizard | no Wall no Crate ... | Crate ] -> [ Crate | ... | Wizard ]

Heroes of Sokoban III uses this more complicated formulation:

    [ > Wizard ] -> [ Wizard > Warpball]
    [ > Warpball | no Crate no Wall] -> [ | > Warpball]
    [ > Warpball | Wall] [Wizard] -> [ | Wall ] [> Wizard]
    [ > Warpball | Crate ] [Wizard] -> [ | Wizard ] [Crate]
    [ > Warpball] [Wizard] -> [] [> Wizard]

This extension is backwards-compatible; an ellipsis pattern "A|B ...|C"
matches "A|C", "A|B|C", "A|B|B|C", etc., so if B is empty, the pattern
naturally matches "A|C", "A||C", "A|||C", etc.