Added testsAutoAdvanceLevel flag.
Sometimes, you might want to write a unit test that stays at the winning state before progressing to the next level, so you can e.g. make sure that the winning state has various desired properties. Or you might want to automatically find a solution to the level without advancing to the next level. This small change makes that possible.