Change how the editor selects glyphs.
For issue #174.

This approach allows selection of glyphs without a matching background but
doesn't include a non-matching background when counting matched bits.

That led to a situation where the editor would sometimes choose a glyph
with a matching background object instead of one with a matching foreground,
so we assign foreground objects more weight.

This commit licensed as MIT/X11.