Print GIF download links to the console.
Chrome blocks direct navigations to data: URLs
(https://crbug.com/594215), which causes generating a GIF to fail (it
just opens a blank window). Fix this by printing a download link to the
GIF in the console instead (which is allowed).

While I'm here, remove the custom base64 implementation in favor of
calling btoa().