add a local storage guard
This bit me when playing a puzzlescript game on itch.io - once the game called
the CHECKPOINT command in its code, the game got all weird (wrong-sized
viewport into the level, multiple copies of the level visible at once, and
grid lines between level squares) because the code threw an exception:

```
Uncaught DOMException: Failed to read the 'localStorage' property from 'Window': Access is denied for this document.
```

I'm pretty sure this is because I have "block 3rd-party cookies" set in my
browser (chrome) and the way itch serves its html games counts as "3rd party".

I tested this code change on a test game I uploaded to itch and it fixed the
problem.