Don't clear message text on restart/cancel
As of 03f6f77, `restart`/`cancel` no longer bypass `message`s, so
clearing the message text causes an exception to be thrown. This
commit no longer clears the message text.

This is one of two options being presented to fix this bug. The
alternative is to revert back to the behaviour of `restart`/`cancel`
bypassing `message`s.