Bypass messages when doing restart/cancel
03f6f77 introduced a couple of changes, one of which was that
`restart`/`cancel` would no longer bypass `message`s. This may have
broken older games and introduced a new bug where it would skip to
the next level instead of `restart`ing or `cancel`ing. This commit
reverts that change (while keeping the change that allows `restart`
and `cancel` to have sound effects).

This is one of two options being presented to fix this bug. The
alternative is to continue bypassing messages with
`restart`/`cancel` and instead don't clear the message text.