Fixed real-time / undo issue
Added diff-to-full undo state conversion on real-time ticks.

(Real-time ticks do not generate new undo states.  This means that undoing a 'diff'-ed state could produce incorrect results if real-time ticks had happened in the intervening interval.  Specifically, diffs are applied to the current state.  In the case that the state had changed due to real-time ticks, the current state would not be the same as the state with respect to which the diff was created.  Applying the diff in such a case could produce an incorrect result.

This commit fixes the issue by converting a diff-type undo on the top of the undo stack to a full undo (using the correct reference state) when handling a real-time tick.)