Added safety measures to line parsing
A line beginning with a comment, a section name, or an equals decorator ('===') would still have the rest of that line parsed as if it were part of a normal line for the current section.  However, since no [section-related] start-of-line processing had occurred for that line, the objects accessed during parsing could be stale, incorrect, or non-existent.  This prevents this by disallowing additional content after section names and equals decorators, and making comments transparent to start-of-line detection.  (It also whites-out start-of-line comments for additional transparency to later processing.)

(Note: This commit also makes PuzzleScript less tolerant to stray right parentheses.)