Delete codeql-analysis.yml
Not sure what if anything it's doing - maybe I'll try it again another time, but because I'm running my own tests locally anyway, it's not likely to contribute much to my workflow. On the other hand, it might help with other people submitting patches, if I got qunit to work somehow. Hmm. Whatever.