Set body margin to 0 on mobile
Most browsers have a default margin of 8px on the `body` element, which was
offsetting the `canvas` element. This was not only pushing the canvas slightly
off screen, but was also affecting the calculation of touch coordinates (which
was most noticable with the PuzzleScript Plus fork because it supports touch
controls).