for task #1001
"you have to define something to be the background" is not reachable because it is pre-empted by the error message "you have to define something to be the background", and in that code if a background isn't found it's just set to a default value. So I just removed the error.